추천 트리 역추적 로직 수정 및 세대 기반 수당 지급 버그 해결

+++ mapp/src/main/java/com/pyj/mapp/dto/PointDto.java
@@ -13,7 +13,7 @@ import java.util.Date;
 @Builder
 public class PointDto {
     private Long pointNo;         // 포인트 적립 번호 (PK)
-    private int userNo;           // 포인트를 받은 사용자 USER_NO
+    private Long userNo;           // 포인트를 받은 사용자 USER_NO
     private Long orderNo;         // 주문번호
     private int amount;           // 적립된 포인트
     private String reason;        // 적립 사유 (예: 구매 적립, 수당 지급 등)
